feat: Add transaction types constants and use them

+++ backend/internal/entity/transaction_type.go
@@ -1,6 +1,6 @@
 package entity
 
 type TransactionType struct {
-	ID   uint   `json:"id" gorm:"primary_key"`
+	ID   byte   `json:"id" gorm:"primary_key"`
 	Name string `json:"name" gorm:"notNull"`
 }
